requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
chromadb
pypdf
sentence-transformers
ollama